separate input fields part of Contact Agent Modal into different module of react element
